deployment/Davis/requirements.txt
mysql-connector-python==8.0.33
requests==2.31.0
google-generativeai==0.3.2
configparser==5.3.0
python-dateutil==2.8.2
langchain-openai